just check if hvpa crd is deployed
