--- pkg/virtualgarden/hvpa_crd_test.go
@@ -1,30 +1,0 @@
-// Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package virtualgarden
-
-import (
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-var _ = Describe("HVPACRD", func() {
-	Describe("#HVPACRDLoading", func() {
-		It("should load the HVPA CRD", func() {
-			crd, err := loadHVPACRD()
-			Expect(err).To(Succeed())
-			Expect(crd.Spec.Names.Kind).To(Equal("Hvpa"))
-		})
-	})
-})
